requirements.txt
boto3==1.9.30
botocore==1.12.30
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Django==2.1.2
django-cors-headers==2.4.0
django-environ==0.4.5
django-polymorphic==2.0.3
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
docutils==0.14
drf-nested-routers==0.91
idna==2.7
Jinja2==2.10
jmespath==0.9.3
livereload==2.6.0
Markdown==3.0.1
MarkupSafe==1.1.0
mkdocs==1.0.4
mkdocs-material==3.1.0
mysqlclient==1.3.13
phonenumberslite==8.10.0
pycrypto==2.6.1
Pygments==2.3.1
PyJWT==1.6.4
pymdown-extensions==6.0
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
requests==2.20.1
s3transfer==0.1.13
six==1.11.0
sqlparse==0.2.4
tornado==5.1.1
urllib3==1.24
watchtower==0.5.4